Rfc: 폴더/파일 구조 수정
