+++ src/app/map/page.tsx
@@ -1,0 +1,7 @@
+import MapContainer from "./_components/MapContainer";
+
+const MapPage = () => {
+  return <MapContainer />;
+};
+
+export default MapPage;
